[FATFS] bool and unsigned macros implementation
previously we have used TRUE/FALSE macros both as boolean and unsigned int
Now BTRUE/BFALSE macros gets used for boolean implementation and UTRUE/UFALSE macros for the unsigned int implementation

+++ packages/ti/fs/fatfs/example/console/src/fs_shell_app_utils.h
@@ -158,7 +158,7 @@ typedef struct fsShellAppUtilsCmdEntry
 /* ========================================================================== */
 /*                            Global Variables Declarations                   */
 /* ========================================================================== */
-extern volatile uint8_t fs_is_media_inserted;
+extern volatile uint32_t fs_is_media_inserted;
 
 /* ========================================================================== */
 /*                          Function Declarations                             */
